feat: move ethereum test configurations in autonomy.test_tools.configuration
