+++ tests/test_agents/test_counter.py
@@ -31,7 +31,7 @@ import pytest
 import requests
 from aea.test_tools.test_cases import AEATestCaseMany
 
-from autonomy.constants import DEFAULT_REQUESTS_TIMEOUT
+from autonomy.test_tools.configurations import DEFAULT_REQUESTS_TIMEOUT
 
 from tests.conftest import ANY_ADDRESS
 from tests.fixture_helpers import UseTendermint
