+++ tests/test_connections/test_ledger/test_ledger_api.py
@@ -41,6 +41,7 @@ from aea.mail.base import Envelope, Message
 from aea.protocols.dialogue.base import Dialogue as BaseDialogue
 from aea_ledger_ethereum import EthereumCrypto
 
+from autonomy.test_tools.configurations import ETHEREUM_KEY_DEPLOYER
 from autonomy.test_tools.docker.base import skip_docker_tests
 from autonomy.test_tools.docker.ganache import DEFAULT_GANACHE_CHAIN_ID
 
@@ -54,8 +55,6 @@ from packages.valory.protocols.ledger_api.dialogues import (
     LedgerApiDialogues as BaseLedgerApiDialogues,
 )
 from packages.valory.protocols.ledger_api.message import LedgerApiMessage
-
-from tests.conftest import ETHEREUM_KEY_DEPLOYER
 
 
 SOME_SKILL_ID = "some/skill:0.1.0"
