+++ tests/test_contracts/test_gnosis_safe/test_contract.py
@@ -37,6 +37,11 @@ from web3.eth import Eth
 from web3.exceptions import SolidityError
 from web3.types import TxData
 
+from autonomy.test_tools.configurations import (
+    ETHEREUM_KEY_PATH_1,
+    ETHEREUM_KEY_PATH_2,
+    ETHEREUM_KEY_PATH_3,
+)
 from autonomy.test_tools.docker.base import skip_docker_tests
 
 from packages.valory.contracts.gnosis_safe.contract import (
@@ -44,12 +49,7 @@ from packages.valory.contracts.gnosis_safe.contract import (
     SAFE_CONTRACT,
 )
 
-from tests.conftest import (
-    ETHEREUM_KEY_PATH_1,
-    ETHEREUM_KEY_PATH_2,
-    ETHEREUM_KEY_PATH_3,
-    ROOT_DIR,
-)
+from tests.conftest import ROOT_DIR
 from tests.helpers.contracts import get_register_contract
 from tests.test_contracts.base import (
     BaseGanacheContractTest,
