+++ tests/test_contracts/test_offchain_aggregator/test_contract.py
@@ -28,13 +28,14 @@ from aea.crypto.registries import crypto_registry
 from aea_ledger_ethereum import EthereumCrypto
 from web3 import Web3
 
+from autonomy.test_tools.configurations import ETHEREUM_KEY_PATH_1
 from autonomy.test_tools.docker.base import skip_docker_tests
 
 from packages.valory.contracts.offchain_aggregator.contract import (
     OffchainAggregatorContract,
 )
 
-from tests.conftest import ETHEREUM_KEY_PATH_1, ROOT_DIR
+from tests.conftest import ROOT_DIR
 from tests.test_contracts.base import BaseGanacheContractTest
